[refactor] 디렉토리 수정 및 import 수정
