+++ src/components/mainPageComponent/typingComponents/categorySelection.tsx
